Reorder empty column code
Moved the empty column creation code to the start of the calc_cost function.

This will allow the rxn_data_setup method to act as a "reset" for the fulldata DataFrame. In addition, the calculated value columns will be "wiped out" before each cost calculation, which should hopefully minimize bad calculations because of repeated recalculations using the wrong data in these columns.