Add OPEX to material cost
The OPEX gets added to the returned cost of the material from the rxn_cost function.

This has some other consequences that I haven't fully thought through, and may require a rewrite. For example, the % RM cost/kg prod changes because the cost of the final molecule is more than the cost of RM only. I think this is okay as is, but I need to check.