Create method to get GSheet values
This abstracts some code that was used to get values from a Google Sheet.

This was used in a couple of other methods, so it was replicated unnecessarily. Also, I needed to do a type check for strings/int, because that changes the call to getting a particular worksheet from the Google Sheet.