Remove DF views in cost calc
The _rxn_cost algorithm was using a mixture of slices and views into the full DataFrame, which was causing problems.

Reworked the code so that a slice copy is used for the major portion of the calculations. This slice copy is reset into the full DataFrame at the end. In a way, this makes the code a little easier to follow.