Set the cost, fix for non-excel cost as well
Set the cost of the product. Needed a new dynamic cost column, which I can set to a cell instead of a value. These cells will need to replace the values in the "Cost" column upon export. Also need to divide the "RM cost/kg rxn" by the number of kg product. Also changed the reference to the sub-rxn costs to be these Cost values and not the RM cost/kg rxn values, which is numerically equivalent but makes more sense for the exported sheets.

There was also an error in the non-Excel cost setting, which was not referenced to the slice of the full DataFrame.