Handle Amount column, convert to equiv
If an Amount column is present, convert the values to equivalents

This required some modification to the `rxn_data_setup` method. First, keep the "Amount" column from the rxn file, if present. Then use that column to calculate equivalents. NOTE!!!! There are certainly some corner cases that will not be captured here.