Concat fulldata and pmi
This new code combines the fulldata and pmi in a way that preserves the fulldata order

In the old code, the sort function was changing the order of the compounds from how they were input into the spreadsheet. There is still a sorting based on the "Step" column. That will be tricky to fix, and will need to be changed in the PMI DataFrame creation as well, if anyone cares. This gives the correct output for the `results` and `excel_save` methods