Clean up results methods
Clean up results methods for all classes

To make this function work seamlessly with all classes, I needed to pull out the DataFrame prep portion of the code and make this into a new private method _prep_results. This allowed me to clean up some of the results code in both the WebApp and Excel classes.