Update missing MW error
Move this error check earlier and add a new message.

The old error messages were not printing correctly in the new Streamlit web app. This doubles up the printing of the error, but it works in all apps. I moved this particular error check earlier because it is kinda common and difficult to fix.