Create a CoreCost class
This contains only the methods that are necessary for the route costing/excel save

Right now, these are just pulled from the old Excel Cost object. Just need to make sure these work as is. The only inputs to this are going to be DataFrames and the only outputs will be DataFrames, which gives it a uniform io structure