Create frontends and helper files
Moved appropriate classes to each file

The Excel/Colab/WebApp classes go into the frontend; the value modifying/scanning/plotting functions are moved to the helpers file. I needed to make sure that all of the proper imports are moved to the appropriate places. Tested with Excel, and it works okay. Not tested in Colab or WebApp. THese need to be updated