Update _sanity_check method
This method had some errors to be addressed, add a kwarg for DF display

The web app only shows the "ValueError" statement. The old statements were not helpful. Now the error message is displayed a couple of times. Also added a flag so that DataFrame display of the "bad" inputs can be limited, which is good for the App as it logs errors.