Test `CoreCalc.calc_cost` method
First pytest on the core cost calculation method.

As the first commit, this required a lot of changes. Created a `tests` folder for all test files, and a `tests\data` folder for data files to import.  Put two clean Excel files into that folder. The first test file only checks the "calc_cost" function of the Core class. Uses Numpy testing functions, which are designed for numerical values. Updated pyproject.toml to make it easier to create testing and dev environments.